Move strategies back onto notice class
